Rename component file to be more accurate
