+++ x-pack/plugins/security_solution/public/detections/pages/detection_engine/detection_engine_no_index.test.tsx
@@ -7,7 +7,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { DetectionEngineNoIndex } from './detection_engine_no_signal_index';
+import { DetectionEngineNoIndex } from './detection_engine_no_index';
 jest.mock('../../../common/lib/kibana');
 
 describe('DetectionEngineNoIndex', () => {
